requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
matplotlib==3.8.2
plotly==5.18.0
APScheduler==3.10.4
PyYAML==6.0.1
webdriver-manager==4.0.1
lxml==4.9.3
